create exchanges module
