+++ commands/base_commands/overrides.py
@@ -24,8 +24,6 @@ from evennia.utils.utils import (make_iter, crop, time_format, variable_from_mod
 from server.utils import arx_utils, prettytable
 from server.utils.exceptions import CommandError
 from commands.base import ArxCommand, ArxPlayerCommand
-from world.dominion.models import CraftingMaterials
-
 
 AT_SEARCH_RESULT = variable_from_module(*settings.SEARCH_AT_RESULT.rsplit('.', 1))
 _DEFAULT_WIDTH = settings.CLIENT_DEFAULT_WIDTH
@@ -371,148 +369,6 @@ class CmdDrop(ArxCommand):
             obj.move_to(caller.location, quiet=True)
             # Call the object script's at_drop() method.
             obj.at_drop(caller)
-
-
-class CmdGive(ArxCommand):
-    """
-    give away things
-
-    Usage:
-      give <inventory obj> = <target>
-      give <inventory obj> to <target>
-      give <amount> silver to <target>
-      give/mats <type>,<amount> to <target>
-      give/resource <type>,<amount> to <target>
-
-    Gives an items from your inventory to another character,
-    placing it in their inventory. give/resource does not require
-    you to be in the same room.
-    """
-    key = "give"
-    locks = "cmd:all()"
-
-    # noinspection PyAttributeOutsideInit
-    def func(self):
-        """Implement give"""
-
-        caller = self.caller
-        to_give = None
-        if not self.args:
-            caller.msg("Usage: give <inventory object> = <target>")
-            return
-        if not self.rhs:
-            arglist = self.args.split(" to ")
-            if len(arglist) < 2:
-                caller.msg("Usage: give <inventory object> to <target>")
-                return
-            self.lhs, self.rhs = arglist[0], arglist[1]
-        if "resource" in self.switches:
-            player = caller.player.search(self.rhs)
-            if not player:
-                return
-            target = player.char_ob
-        else:
-            target = caller.search(self.rhs)
-        if not target:
-            return
-        if target == caller:
-            caller.msg("You cannot give things to yourself.")
-            return
-        if not target.player_ob:
-            self.msg("You cannot give anything to them. Use 'put' instead.")
-            return
-        if "mats" in self.switches:
-            lhslist = self.lhs.split(",")
-            try:
-                mat = caller.player_ob.Dominion.assets.materials.get(type__name__iexact=lhslist[0])
-                amount = int(lhslist[1])
-                if amount < 1:
-                    raise ValueError
-            except (IndexError, ValueError):
-                caller.msg("Invalid syntax.")
-                return
-            except CraftingMaterials.DoesNotExist:
-                caller.msg("No materials by that name.")
-                return
-            if mat.amount < amount:
-                caller.msg("Not enough materials.")
-                return
-            try:
-                tmat = target.player_ob.Dominion.assets.materials.get(type=mat.type)
-            except CraftingMaterials.DoesNotExist:
-                tmat = target.player_ob.Dominion.assets.materials.create(type=mat.type)
-            mat.amount -= amount
-            tmat.amount += amount
-            mat.save()
-            tmat.save()
-            caller.msg("You give %s %s to %s." % (amount, mat.type, target))
-            target.msg("%s gives %s %s to you." % (caller, amount, mat.type))
-            return
-        if "resource" in self.switches:
-            rtypes = ("economic", "social", "military")
-            lhslist = self.lhs.split(",")
-            try:
-                rtype = lhslist[0].lower()
-                amount = int(lhslist[1])
-                if amount < 1:
-                    raise ValueError
-            except (IndexError, ValueError):
-                caller.msg("Invalid syntax.")
-                return
-            if rtype not in rtypes:
-                caller.msg("Type must be in %s." % ", ".join(rtypes))
-                return
-            cres = getattr(caller.player_ob.Dominion.assets, rtype)
-            if cres < amount:
-                caller.msg("You do not have enough %s resources." % rtype)
-                return
-            tres = getattr(target.player_ob.Dominion.assets, rtype)
-            cres -= amount
-            tres += amount
-            setattr(target.player_ob.Dominion.assets, rtype, tres)
-            setattr(caller.player_ob.Dominion.assets, rtype, cres)
-            target.player_ob.Dominion.assets.save()
-            caller.player_ob.Dominion.assets.save()
-            caller.msg("You give %s %s resources to %s." % (amount, rtype, target))
-            target.player_ob.inform("%s has given %s %s resources to you." % (caller, amount, rtype),
-                                    category="Resources")
-            return
-        if args_are_currency(self.lhs):
-            arglist = self.lhs.split()
-            val = round(float(arglist[0]), 2)
-            if val <= 0:
-                self.msg("Amount must be positive.")
-                return
-            currency = round(float(caller.db.currency or 0), 2)
-            if val > currency:
-                caller.msg("You do not have that much money to give.")
-                return
-            caller.pay_money(val, target)
-            caller.msg("You give coins worth %s silver pieces to %s." % (val, target))
-            target.msg("%s has given you coins worth %s silver pieces." % (caller, val))
-            return
-        # if we didn't find a match in currency that we're giving
-        if not to_give:
-            to_give = caller.search(self.lhs)
-        if not (to_give and target):
-            return
-        if target == caller:
-            caller.msg("You keep %s to yourself." % to_give.key)
-            to_give.at_get(caller)
-            return
-        if not to_give.location == caller:
-            caller.msg("You are not holding %s." % to_give.key)
-            return
-        if not check_volume(to_give, target, quiet=True):
-            caller.msg("%s can't hold %s." % (target.name, to_give.name))
-            return
-        if not to_give.at_before_move(target, caller=caller):
-            return
-        # give object
-        to_give.move_to(target, quiet=True)
-        caller.msg("You give %s to %s." % (to_give.key, target))
-        target.msg("%s gives you %s." % (caller, to_give.key))
-        to_give.at_get(target)
 
 
 class CmdEmit(ArxCommand):
